move debug cmd tests to own module
